s400: Warn and do not report level/SNR if NaN
The S400 returns strings "< 70" and "< 3" if the signal level and CNR
are below -70 dBm and 3 dB, respectively. In this case, because the
actual numbers are uncertain, it's better not to include them in the
report as NaN. For instance, the report-receiving side (like the
Monitoring API) could reject the NaN values. Hence, instead of sending
NaNs, log warnings so that the user knows what's happening.