satip/s400: Support carrier freq offset correction
It is not unusual to have an LNB with LO frequency offset exceeding the
regular DVB-S2's coarse frequency offset estimation range. Hence, in
some cases, it is preferable to tune to a center frequency with a fixed
offset relative to the nominal value. This patch adds such an option
(named --freq-corr) on the satip and standalone receiver commands. With
the sat-ip receiver, the frequency correction has to be specified every
time on the launch command (blocksat-cli sat-ip). Meanwhile, with the
standalone demodulator (Novra S400), it has to be specified once on the
configuration step (blocksat-cli standalone cfg).