ci: Split and update the fedora and centos jobs
Separate the centos and fedora jobs to install epel-release on the
centos jobs. Install also gcc and python3-devel onc CentOS so that the
zfec dependency can be built for python 3.6. The latest zfec does not
have a wheel available for the py36 used on centos8. Furthermore, fix
the yum repos on centos 8 to ensure "dnf install" works despite the
changes due to centos8 reaching EOL.

On the Fedora jobs, test under fc34 and fc35, the two latest releases as
of this time.