deps: Add function to check the required apps
The names of the applications required to run each receiver are distinct
to the names of the binary packages that contain them. When launching
the receiver, we want to check the applications, not the binary
packages. Hence, add a utility function that can check all applications
for a given receiver type. Use it where applicable, like the Sat-IP
client. However, preserver the more custom verifications that check
applications individually depending on the given configuration, such as
the SDR module, which does not check tsp when the command disables it.