Use apt instead of apt-get
Previously, 'apt' was used in most calls except the installation (where
apt-get was used). For consistency, use the same call everywhere. This
is useful when, e.g., pkexec is used under the hood to assign permission
for a specific application with caching. Once apt has permissions,
further calls to it won't request permission again.