Support reports authenticated via password
Add support for password-authenticated reports sent to the Blockstream
Satellite Monitoring API as an alternative to GPG-authenticated reports.
Create a new password automatically when necessary and use it whenever
available as the preferred authentication method instead of the GPG
approach, given this method has low CPU utilization compared to
GPG-signing reports periodically.

When the password is generated for the account, save it locally on an
encrypted file that can be decrypted using the GPG key. Also, flag the
monitoring password availability on the new monitoring.has_password
field of the JSON config file.

Also, the previous implementation had a debugging log for the entire
reporting request's payload, including the authentication data. With the
new password-based mechanism, this log would include the password, which
is part of the payload sent over https. To avoid logging the password,
this debugging log now excludes the auth info.