Fix report passwords reused by multiple configs
The previous implementation would fail for multiple configurations
reporting using the same GPG keypair. They would end up saving the
password file with the same name and overwriting one another. The
passphrase has to be different for each configuration even if the same
GPG keypair is used. Hence, include the configuration name on the
password file to ensure it is unique.