api: Use /admin endpoints when using TLS key/cert
Use the new admin endpoints introduced on the API with multichannel
support. The admin endpoints provide read/write/delete permissions (to
the admin) even on the endpoints that don't allow users to send
get/post/delete requests.

Switch to the /admin endpoint automatically when running with a TLS
certificate given that the cert/key is intended to be used for admin
authentication.