demo-rx: Drop --gossip and --btc-src options
With the new multichannel API server implementation, all that it takes
to run the demo-rx app on the gossip and btc-src channels is to specify
the --channel parameter. The previous --gossip and --btc-src options
would only override the channel argument, which is unnecessary.
Previously, they were used to override both the channel and the server.