tests: Use a common fixture to create test dirs
Some tests create the configuration directory and/or GPG directory.
Previously, they would be created at fixed paths, which could cause
conflicts with tests running in parallel. This patch defines a common
helper fixture called TestEnv that can be reused in tests to handle the
creation and deletion of the config and GPG directories.