demo-rx: Add option to map sockets to regions
Previously, all sockets in use by the DemoRx class would send all
orders, as long as the order had at least of the regions specified by
the demo-rx's regions set. This patch adds an option to assign each
socket to a single region. With that, the sockets will only send the
messages covered by their assigned regions.

For instance, if the demo-rx app is launched with --interfaces en0 en1
and --regions 0 1, then interface en0 is assigned to region 0 and
interface en1 to region 1. If an order comes for region 1 only, it will
only be sent over interface en1.