monitoring-api: Refactor password reset mechanism
The previous implementation had option --bs-mon-reset-pwd to reset the
password in runtime, when launching the receiver. This patch refactors
removes this option while creating a new command to manage anything
related to the monitoring API. Now, the password can be reset by command
"blocksat-cli reporting password". There is also a command "reporting
info" to show information of the current registration. Both commands can
be executed with no need for the receiver to be running.