monitoring-api: Support report to alternate server
The previous implementation enabled reports to the Blockstream
Monitoring API every time the --report-dest was set to the Monitoring
API URL, which was the default. However, this approach prevented testing
with an alternate Monitoring API server (e.g., local or staging). Also,
it did not facilitate unit testing with a mock server URL.

This patch makes sure the Monitoring API server URL can be defined on
construction of the BsMonitoring object so that reports can be sent to
alternate Monitoring API servers and unit tests don't need to interact
with the production server.