gpg: Drop dependency on blocksat key import order
Previously, the get_default_public_key function would only work as
intended if the blocksat pubkey was imported after creating the user's
keypair. However, this approach was error prone in case the key creation
failed and the blocksat pubkey was imported anyway. Update the
implementation so that get_default_public_key no longer depends on
importing the blocksat pubkey with the right order.