fec: Make FEC support optional
The FEC support is only used by the API sender and listener applications
in specific circumstances (optional on the send side). Since zfec is not
available on Fedora as a package (like the python3-zfec package
available on Ubuntu), having it on the requirements would prevent
building an RPM package. Make it optional to allow building RPM
packages.