firewall: Refactor to use classes and factory
Implement a standard firewall base class with two abstract methods: one
for verification of the current configuration, the other to apply the
configuration. Implement two classes based on this abstract class for
the two supported firewall implementations, iptables and firewalld.

Also, extend the coverage of unit tests using the new classes.