firewall: Use no-check on firewall-cmd query
Previously, when the firewall-cmd query commands did not find any rule,
the ProcessRunner would raise a CalledProcessError and break the code.
That is because the firewall-cmd exits with code 1 when no rule is
found. Since the logic relies on the output captured via  stdout, do not
check the exit code from the process and rely on stdout only.