cfg: Patch channels conf file on updates
When the satellite information is changed on a CLI update and the user
has a Linux USB setup, patch the channels.conf file automatically. With
that, the user won't need to rerun the "blocksat-cli cfg channel"
command. The user can simply relaunch the USB receiver and the new
channel information will be ready.