ci: Install dependencies before buildx build
The Makefile now builds the blocksat CLI manpage before building the
Blocksat Host container image. However, help2man (executed for the
manpage build) will not work if the package dependencies are not
installed, as it needs to run ./blocksat-cli --help successfully. As a
workaround, install the package (run "make install") before running
"make buildx" so that the dependencies are installed first.