standalone: Use pyasyncore for PySNMP on Py3.12
The current PySNMP version relies on the asyncore module, which was
removed in Python 3.12. While a new version replacing asyncore with
asyncio (or something else) is not yet available, we can use the
pyasyncore module as a drop-in replacement. In the future, once a proper
fix is included in PySNMP, we can remove the pyasyncore dependency.